+++ global_helpers/panther_okta_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_okta_helpers.py
+GlobalID: "panther_okta_helpers"
+Description: >
+  Used to define global helpers and variables for Okta events.
