+++ rules/okta_rules/okta_org2org_creation_modification.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import okta_alert_context
+from panther_okta_helpers import okta_alert_context
 
 APP_LIFECYCLE_EVENTS = (
     "application.lifecycle.update",
