+++ rules/okta_rules/okta_support_reset.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import okta_alert_context
+from panther_okta_helpers import okta_alert_context
 
 OKTA_SUPPORT_RESET_EVENTS = [
     "user.account.reset_password",
