+++ rules/okta_rules/okta_user_mfa_reset_all.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import okta_alert_context
+from panther_okta_helpers import okta_alert_context
 
 
 def rule(event):
